requirements.txt
airtest==1.3.1
numpy==1.24.4
opencv_contrib_python==4.6.0.66
openpyxl==3.1.2
pocoui==1.0.89
PyAutoGUI==0.9.54
